Allow users to provide arbitrary tls.Config for WebClient
This commit replaces the previous WebClient options `InsecureTLS` and
`RootCAs` with a new option `TLSConfig` which can be used to pass an
arbitrary tls.Config object which is configured in the WebClient
`http.Transport`.

If the TLSConfig option is `nil`, Go's default configuration is used.

We set `ForceAttemptHTTP2` on the `http.Transport` object to true to
ensure the HTTP/2 upgrade is attempted even when a custom TLS
configuration is passed.

This change allows users of go-icinga2-client to configure the TLS
connection by directly providing a tls.Config which contains all the
configuration they require.